+++ mdp_client.c
@@ -119,7 +119,7 @@ int overlay_mdp_client_init()
       if(config.debug.io) DEBUGF("MDP client socket name='%s'",overlay_mdp_client_socket_path);
     }
     if (overlay_mdp_client_socket_path_len > sizeof(name.sun_path) - 1)
-      FATALF("MDP socket path too long (%d > %d)", overlay_mdp_client_socket_path_len, sizeof(name.sun_path) - 1);
+      FATALF("MDP socket path too long (%d > %d)", overlay_mdp_client_socket_path_len, (int)sizeof(name.sun_path) - 1);
     
     bcopy(overlay_mdp_client_socket_path,name.sun_path,
 	  overlay_mdp_client_socket_path_len);
