+++ overlay_interface.c
@@ -881,13 +881,13 @@ overlay_broadcast_ensemble(overlay_interface *interface,
 	if (fsize == -1)
 	  return WHY_perror("lseek");
 	if (config.debug.overlayinterfaces)
-	  DEBUGF("Write to interface %s at offset=%d", interface->name, fsize);
+	  DEBUGF("Write to interface %s at offset=%"PRId64, interface->name, fsize);
       }
       ssize_t nwrite = write(interface->alarm.poll.fd, &packet, sizeof(packet));
       if (nwrite == -1)
 	return WHY_perror("write");
       if (nwrite != sizeof(packet))
-	return WHYF("only wrote %lld of %lld bytes", nwrite, sizeof(packet));
+	return WHYF("only wrote %d of %d bytes", (int)nwrite, (int)sizeof(packet));
       return 0;
     }
